Move remaining PNL functions to wendigo_pnl.c
